fix: 센서 데이터 파이프라인을 raw uint32로 변경 (float 변환 제거)
센서 데이터는 정수 원본값(ADC count, mm 등)이므로 float 변환 없이
uint32 그대로 저장/전송해야 함. 기존에는 Uint32ToFloat(IEEE 754 비트캐스트)
를 적용하여 hand_log.csv에 0으로 기록되는 버그가 있었음.

변경 사항:
- HandState.sensor_data: float → uint32_t
- LogEntry.hand_sensors: float → uint32_t
- RequestSensorRead: DecodeSensorResponse → DecodeSensorResponseRaw
- hand_failure_detector: 센서 비교 타입 uint32_t로 통일
- rt_controller_node: sim 센서 수신 타입 uint32_t로 변경
- ExtractSensorValues(float): 원복 (legacy 유지, 프로덕션 미사용)

+++ ur5e_rt_base/include/ur5e_rt_base/logging/log_buffer.hpp
@@ -76,7 +76,7 @@ struct LogEntry {
   // 카테고리 2: Current State
   std::array<float, kNumHandMotors>  hand_actual_positions{};
   std::array<float, kNumHandMotors>  hand_actual_velocities{};
-  std::array<float, kMaxHandSensors> hand_sensors{};           // 최대 kMaxFingertips × 11
+  std::array<uint32_t, kMaxHandSensors> hand_sensors{};         // 최대 kMaxFingertips × 11 (raw uint32)
   int num_fingertips{kDefaultNumFingertips};                   // 실제 사용 fingertip 수
 
   // 카테고리 3: Control Command
